requirements.txt
face_recognition==1.3.0
firebase_admin==6.5.0
Flask==2.2.5
Flask_Cors==5.0.0
Flask_SocketIO==5.3.6
numpy==2.1.3
Pillow==11.0.0
protobuf==5.28.3
